Default numpy array type is incorrect

+++ src/nidigital/system_tests/test_system_nidigital.py
@@ -144,8 +144,8 @@ def test_source_waveform_parallel_site_unique(multi_instrument_session, source_w
             0: array.array('L', [i for i in reversed(range(num_samples))])}
     elif source_waveform_type == numpy.array:
         source_waveform = {
-            1: numpy.array([i for i in range(num_samples)]),
-            0: numpy.array([i for i in reversed(range(num_samples))])}
+            1: numpy.array([i for i in range(num_samples)], dtype=numpy.uint32),
+            0: numpy.array([i for i in reversed(range(num_samples))], dtype=numpy.uint32)}
     elif source_waveform_type == list:
         source_waveform = {
             1: [i for i in range(num_samples)],
